requirements.txt
streamlit
langchain_groq 
langchain_community 
pypdf 
PyMupdf
sentence_transformers
python-dotenv
chromadb
transformers
langchain-chroma
langchain-huggingface

fastapi
uvicorn[standard]
langchain-core
langchain-groq

chromadb
sentence-transformers
PyPDF
pydantic
loguru
aiofiles